requirements.txt
absl-py==0.7.1
astor==0.8.0
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.2
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.20.1
h5py==2.9.0
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
msgpack==0.6.1
numpy==1.17.0
Pillow==6.1.0
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.25.2
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
